requirements.txt
pytest==8.3.3
pytest-asyncio==0.25.1
grpcio==1.69.0
grpcio-tools==1.69.0
allure-pytest==2.13.5
allure-python-commons~=2.13.5